Separate reaction generation code as a separate class.
